requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
tqdm
tensorboard
blitz-bayesian-pytorch
